requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openai>=1.3.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
langchain>=0.0.350
langchain-community>=0.0.10
numpy>=1.24.0
httpx>=0.25.0
pytest>=7.4.0
requests>=2.31.0
